Update ShinyGUIClass.R
Replacing NA with 0 circumvents an error that arises when using Inf on the timer: "Error: missing value where TRUE/FALSE needed". As far as I can tell, the error arises because server.R line 43:

if(!is.null(item) && .MCE[[sessionName]]$shinyGUI$timer[item] > 0)

This if statement cannot take NA as an input and therefore throws an error.